test(deducer): add dependency analysis tests
