+++ components/deducers/static/src/imports.test.ts
@@ -1,33 +1,7 @@
+import ts from "typescript";
 import { test, expect } from "vitest";
 import { ImportType, extractImportElements } from "./imports";
-
-import ts from "typescript";
-
-function genInlineSourceFile(sourceCode: string): ts.SourceFile {
-  const fileName = "inline.ts";
-  const host: ts.LanguageServiceHost = {
-    getCompilationSettings: () => ({
-      target: ts.ScriptTarget.ES2022,
-      module: ts.ModuleKind.CommonJS,
-    }),
-    getCurrentDirectory: () => process.cwd(),
-    getDefaultLibFileName: (options) => ts.getDefaultLibFilePath(options),
-    getScriptFileNames: () => [fileName],
-    getScriptVersion: () => "0",
-    getScriptSnapshot: (name) =>
-      name === fileName ? ts.ScriptSnapshot.fromString(sourceCode) : undefined,
-    readFile: () => undefined,
-    fileExists: () => true,
-  };
-  const service = ts.createLanguageService(host);
-
-  const program = service.getProgram();
-  expect(program).toBeDefined();
-
-  const sourceFile = program!.getSourceFile(fileName);
-  expect(sourceFile).toBeDefined();
-  return sourceFile!;
-}
+import { genInlineSourceFile } from "./utils";
 
 test("case: direct import", () => {
   const importStat = `import "fs";`;
